move garbage collector logic to index-management
