+++ quickwit/quickwit-janitor/src/actors/garbage_collector.rs
@@ -23,15 +23,14 @@ use std::sync::Arc;
 use std::time::Duration;
 
 use async_trait::async_trait;
-use futures::{stream, StreamExt};
+use futures::{stream, Future, StreamExt};
 use itertools::Itertools;
 use quickwit_actors::{Actor, ActorContext, Handler};
+use quickwit_index_management::{run_garbage_collect, ExternalCallDelegate};
 use quickwit_metastore::Metastore;
 use quickwit_storage::StorageResolver;
 use serde::Serialize;
 use tracing::{error, info};
-
-use crate::garbage_collection::run_garbage_collect;
 
 const RUN_INTERVAL: Duration = Duration::from_secs(10 * 60); // 10 minutes
 
@@ -78,6 +77,19 @@ pub struct GarbageCollector {
     counters: GarbageCollectorCounters,
 }
 
+#[derive(Clone, Copy)]
+struct ActorAwareCallDelegate<'a> {
+    ctx: &'a ActorContext<GarbageCollector>,
+}
+
+#[async_trait]
+impl ExternalCallDelegate for ActorAwareCallDelegate<'_> {
+    async fn delegate<Fut, T>(&self, future: Fut) -> T
+    where Fut: Future<Output = T> + Send {
+        self.ctx.protect_future(future).await
+    }
+}
+
 impl GarbageCollector {
     pub fn new(metastore: Arc<dyn Metastore>, storage_resolver: StorageResolver) -> Self {
         Self {
@@ -114,6 +126,7 @@ impl GarbageCollector {
                     return None;
                 }
             };
+            let ctx_aware_delegate = ActorAwareCallDelegate { ctx };
             let index_uid = index.index_uid;
             let gc_res = run_garbage_collect(
                 index_uid.clone(),
@@ -122,7 +135,7 @@ impl GarbageCollector {
                 STAGED_GRACE_PERIOD,
                 DELETION_GRACE_PERIOD,
                 false,
-                Some(ctx),
+                ctx_aware_delegate,
             ).await;
             Some((index_uid, gc_res))
         }}).buffer_unordered(MAX_CONCURRENT_GC_TASKS);
@@ -320,7 +333,7 @@ mod tests {
             STAGED_GRACE_PERIOD,
             DELETION_GRACE_PERIOD,
             false,
-            None,
+            (),
         )
         .await;
         assert!(result.is_ok());
